% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie_group_report.R
\name{bowtie_group_report}
\alias{bowtie_group_report}
\title{Creates a table from the alignment reports from bowtie, containing information from a
all .log files in a directory.}
\usage{
bowtie_group_report(path = NULL, suffix = ".log", samples = NULL)
}
\arguments{
\item{path}{The path to the directory containing the log files from the bowtie
alignment.}

\item{suffix}{The suffix of the report files created by the code similar to:
bowtie2 -x ref -1 sample1_R1.fq -2 sample1_R2.fq -S sample1.sam 2> sample1.log
In this case it would be .log. Default ".log"}

\item{samples}{A vector containing the names of the samples, should be the same
length as list.files(path, pattern = suffix), and follow the same order. If NULL,
sample names will be the file name without the suffix given.}
}
\value{
data frame of the stats contained in the log file from bowtie for all samples
in the directory.
}
\description{
Creates a table from the alignment reports from bowtie, containing information from a
all .log files in a directory.
}
\examples{
dir <- file.path(system.file(paste0("extdata/testdata/Bowtie/"), package = "SeqReport"))
bowtie_group_report(path = dir)
bowtie_group_report(path = dir, samples = c("sp1", "sp2", "sp3"))

}
