% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie_heatmap.R
\name{bowtie_heatmap}
\alias{bowtie_heatmap}
\title{Bowtie alignment heatmap}
\usage{
bowtie_heatmap(
  dt_bowtie = NULL,
  total_reads = TRUE,
  alignment_rate = TRUE,
  headers = "all",
  samples = NULL,
  value = "raw",
  filter_samples = NULL,
  barplot = FALSE
)
}
\arguments{
\item{dt_bowtie}{data.frame or tibbles produced by the function bowtie_group_report.}

\item{total_reads}{if you would like to display the total number of reads for each sample.
Default = TRUE.}

\item{alignment_rate}{if you would like to display the alignment rate for each sample.
Default = TRUE.}

\item{headers}{if only the main statistical headers or all headers should be used.
the main statistical headers are the first 4 stats of the bowtie report. Can be "main" or
"all". Default = "all".}

\item{samples}{a vector of the same length as the number of rows in dt_bowtie, to be used to name
the samples. If not present the name of the samples will be the the "Sample" column
from the dt_bowtie table.}

\item{value}{which value should be displayed in the heatmap. The percentages or the raw number of reads.
Possibilities: "raw" or "percent". Default = "raw".}

\item{filter_samples}{vector with names of samples to be excluded from visualization.
if samples parameter is provided, the samples to filter should be present in the samples parameter.}

\item{barplot}{logical, default = FALSE. if the total reads should be displayed as a bar plot.
Only works if the total_reads is set to TRUE.}
}
\value{
a Heatmap object.
}
\description{
produces a heatmap of the number of reads aligned in the different
alignment parameterd given by the report from bowtie. It needs to be used with data.frames
or tibbles produced by the function bowtie_group_report.
}
\examples{

dir <- file.path(system.file(paste0("extdata/testdata/Bowtie/"), package = "SeqReport"))
dt_bowtie <- bowtie_group_report(path = dir, samples = c("sp1", "sp2", "sp3"))
bowtie_heatmap(dt_bowtie)

}
