% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie_report_summary.R
\name{bowtie_report_summary}
\alias{bowtie_report_summary}
\title{bowtie_report_summary}
\usage{
bowtie_report_summary(
  path = NULL,
  dt_bowtie = NULL,
  suffix = ".log",
  samples = NULL,
  exclude = NULL
)
}
\arguments{
\item{path}{The path to the directory containing the log files from the bowtie
alignment. Only necessary if no dt_bowtie is given.}

\item{dt_bowtie}{data.frame or tibble resulting from bowtie_group_report}

\item{suffix}{The suffix of the report files created by the code similar to:
bowtie2 -x ref -1 sample1_R1.fq -2 sample1_R2.fq -S sample1.sam 2> sample1.log
In this case it would be .log. Default ".log". Only used if path is not null.}

\item{samples}{A vector containing the names of the samples, should be the same
length as list.files(path, pattern = suffix), and follow the same order. If NULL,
sample names will be the file name without the suffix given. Can only be
used if path is not null.}

\item{exclude}{character vector containing the name of the samples to be excluded
from the summary.}
}
\value{
A printed summary of the bowtie logs.
}
\description{
bowtie_report_summary
}
\examples{
parent_testdata_dir <- file.path(system.file(paste0("extdata/testdata/"),
package = "SeqReport"))
data_table <- readRDS(file.path(parent_testdata_dir, "test_bowtie_no_samples.RDS"))
bowtie_report_summary(dt_bowtie = data_table)


}
