% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwa_heatmap.R
\name{bwa_heatmap}
\alias{bwa_heatmap}
\title{bwa_heatmap}
\usage{
bwa_heatmap(
  dt_bwa = NULL,
  total_reads = TRUE,
  alignment_rate = TRUE,
  samples = NULL,
  value = "raw",
  filter_samples = NULL,
  barplot = FALSE
)
}
\arguments{
\item{dt_bwa}{data.frame or tibbles produced by the function bwa_report_group.}

\item{total_reads}{if you would like to display the total number of reads for each sample.
Default = TRUE.}

\item{alignment_rate}{if you would like to display the alignment rate for each sample.
Default = TRUE.}

\item{samples}{a vector of the same length as the number of rows in dt_bwa, to be used to name
the samples. If not present the name of the samples will be the the "Sample" column
from the dt_bwa table.}

\item{value}{which value should be displayed in the heatmap. The percentages or the raw number of reads.
Possibilities: "raw" or "percent". Default = "raw".}

\item{filter_samples}{vector with names of samples to be excluded from visualization.
if samples parameter is provided, the samples to filter should be present in the samples parameter.}

\item{barplot}{logical, default = FALSE. if the total reads should be displayed as a bar plot.
Only works if the total_reads is set to TRUE.}
}
\value{
a Heatmap object.
}
\description{
bwa_heatmap
}
\examples{

parent_testdata_dir <- file.path(system.file(paste0("extdata/testdata/"),
package = "SeqReport" ))
testdata_dir <- file.path(parent_testdata_dir, "BWA")
dt_bwa_ex <- bwa_report_group(path = testdata_dir, suffix = ".txt",
suffix_to_remove = "_report.txt",
samples = NULL)
bwa_heatmap(dt_bwa = dt_bwa_ex,
total_reads = TRUE,
alignment_rate = TRUE,
samples = NULL,
value = "raw",
filter_samples = NULL,
barplot = FALSE)


}
