% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwa_report_group.R
\name{bwa_report_group}
\alias{bwa_report_group}
\title{bwa_report_group}
\usage{
bwa_report_group(
  path = NULL,
  suffix = ".txt",
  suffix_to_remove = "_merged_panel_nt_report.txt",
  samples = NULL
)
}
\arguments{
\item{path}{path to the directory containing the bwamem report files}

\item{suffix}{suffix of the files containing the bwamem report, usually ".txt"}

\item{suffix_to_remove}{if you want the sample name to be derived from the files
with a suffix removal. default: "_merged_panel_nt_report.txt"}

\item{samples}{a character vector containing the names to replace the samples names
in the same order as the files. default: NULL}
}
\value{
a data.frame object
}
\description{
bwa_report_group
}
