% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwa_report_summary.R
\name{bwa_report_summary}
\alias{bwa_report_summary}
\title{bwa_report_summary}
\usage{
bwa_report_summary(
  path = NULL,
  dt_bwa = NULL,
  suffix = ".txt",
  suffix_to_remove = "_merged_panel_nt_report.txt",
  samples = NULL,
  exclude = NULL
)
}
\arguments{
\item{path}{path to the directory containing the bwamem report files}

\item{dt_bwa}{data.frame or tibbles produced by the function bwa_group_report.}

\item{suffix}{suffix of the files containing the bwamem report, usually ".txt"}

\item{suffix_to_remove}{if you want the sample name to be derived from the files
with a suffix removal. default: "_merged_panel_nt_report.txt"}

\item{samples}{a character vector containing the names to replace the samples names
in the same order as the files. default: NULL}

\item{exclude}{@param exclude character vector containing the name of the samples to be excluded
from the summary.}
}
\value{
A printed message with the summary of the statistical information on the reports.
}
\description{
bwa_report_summary
}
\examples{
parent_testdata_dir <- file.path(system.file(paste0("extdata/testdata/"),
package = "SeqReport" ))
testdata_dir <- file.path(parent_testdata_dir, "BWA")
bwa_report_summary(path = testdata_dir, dt_bwa = NULL,
suffix = ".txt",
suffix_to_remove = "_report.txt",
samples = NULL,
exclude = NULL)


}
