% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc_group_report.R
\name{.filter_info}
\alias{.filter_info}
\title{Internal function - generic function to filter the basic stats columns based on info.}
\usage{
.filter_info(fastqc_table, info = "warnings")
}
\arguments{
\item{info}{which information should be added to the table: "all" All basic stats,
"warnings" the statistics with warning, "warn_fail" the statistics with warning and the
ones with fail, "fails" only the statistics that failed, "pass" the statistics that passed.
default "all".}

\item{col}{which column to apply the function to.}
}
\description{
Internal function - generic function to filter the basic stats columns based on info.
}
\keyword{internal}
