% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie_group_report.R
\name{.import_text_from_log}
\alias{.import_text_from_log}
\title{helper for bowtie_group_report}
\usage{
.import_text_from_log(file)
}
\arguments{
\item{file}{}
}
\value{
datatable of individual files
}
\description{
helper for bowtie_group_report
}
\keyword{internal}
