% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie_heatmap.R
\name{.value_fun}
\alias{.value_fun}
\title{Helper for the bowtie_heatmap function}
\usage{
.value_fun(x)
}
\arguments{
\item{x}{data.frame from function bowtie_group_report.}
}
\value{
a tibble.
}
\description{
extracts the raw value of the data.frames
or tibbles produced by the function bowtie_group_report.
}
\keyword{internal}
