% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc_group_report.R
\name{fastqc_group_report}
\alias{fastqc_group_report}
\title{fastqc_group_report}
\usage{
fastqc_group_report(path = NULL, info = "all", stats = TRUE, samples = NULL)
}
\arguments{
\item{path}{path to the directory that contain one or more .html files from the
FastQC.}

\item{info}{which information should be added to the table: "all" All basic stats,
"warnings" the statistics with warning, "warn_fail" the statistics with warning and the
ones with fail, "fails" only the statistics that failed, "pass" the statistics that passed.
default "all".}

\item{stats}{logical, if the sequence length, total number of sequences, %GC content
and unique overrepresented sequences sources should be added to the table. default: TRUE}

\item{samples}{character vector. Vector containing the names of the samples for the files
to be reported in the directory in the same order as in the directory. default NULL.
If NULL, the basename of the files are going to be used.}
}
\value{
dataframe containing the statistical information from all fastqc.html files in the directory
in path.
}
\description{
This function will get the information from all .html files from the
FastQC evaluation presented in a directory and will combine the information in a
data_table.
}
\examples{
\dontrun{
path <- "path/to/directory/containing/html/files"
table_fastqc <- fastqc_group_report(
  path = path,
  info = all,
  stats = TRUE,
  samples = NULL
)
}

}
