% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc_report_summary.R
\name{fastqc_report_summary}
\alias{fastqc_report_summary}
\title{Fastqc report summary}
\usage{
fastqc_report_summary(fastqc_table = NULL, path = NULL, exclude = NULL)
}
\arguments{
\item{fastqc_table}{table created from the fastqc_group_report. if the table is
not available the directory containing the fastqc .html files can be input on the
path parameter}

\item{path}{if there is no fastqc_table, a path to the directory containing the
fastqc .html files can be added}

\item{exclude}{character vector containing the name of samples to be
excluded in the summary}
}
\value{
a printed summary of the information in the .html, such as:
Average number of reads,
Minimum number of reads (and samples),
Maximum number of reads (and samples),
Average sequence length,
Minimum sequence length (and samples),
Samples with over represented adapter sequences.
Red flags (fastqc flags that failed in at least 50% of the samples)
}
\description{
Fastqc report summary
}
\examples{

parent_testdata_dir <- file.path(system.file(paste0("extdata/testdata/"),
package = "SeqReport"))
data_table <- readRDS(file.path(parent_testdata_dir, "test_fqc_table_no_samples.RDS"))
fastqc_report_summary(fastqc_table = data_table)



}
