% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_html_table.R
\name{read_html_table}
\alias{read_html_table}
\title{reads a html table from file into R}
\usage{
read_html_table(file, info = "stats", transpose = TRUE)
}
\arguments{
\item{file}{path to the html file.}

\item{info}{The table to be extracted: "stats" stats table or "hits" the overrepresented
sequences table. Defaul "stats".}

\item{transpose}{logical, if the table should be transposed or not.
Default = TRUE.}
}
\value{
a tibble from the chosen table from the fastqc html file.
}
\description{
function to get the table from the html file and transpose it.
}
\examples{
\dontrun{
file <- "path/to/file/file.html"
table <- read_html_table(file = file, info = stats, transpose = TRUE)
}

}
