% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc_report_summary.R
\name{red_flags_fun}
\alias{red_flags_fun}
\title{Helper for the fastq_report_summary function}
\usage{
red_flags_fun(fastqc_table)
}
\arguments{
\item{fastqc_table}{table created from the fastqc_group_report.}
}
\value{
a character vector containing the name of the columns.
}
\description{
extracts the columns that have failed flags for more than 50% of the samples.
}
\keyword{internal}
