% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_html_table.R
\name{read_html_table}
\alias{read_html_table}
\title{reads a html table from file into R}
\usage{
read_html_table(file, info = c("stats", "hits"), transpose = TRUE)
}
\arguments{
\item{file}{path to the html file.}

\item{info}{if you want the stats table or the over represented sequences table.}

\item{transpose}{logical, if the table should be transposed or not. Default = TRUE.}
}
\value{
a tibble from the chosen table from the fastqc html file.
}
\description{
helper function to get the table from the html file and transpose it.
}
\examples{
\dontrun{
     file <- path/to/file/file.html
     table <- read_html_table(file = file, info = stats, transpose = TRUE)
}

}
